controller fixes
Not sure what made controller suddenly start working, but I'm not going to look a gift horse in the mouth. I think it just wasn't connecting correctly or something. So I was able to remove the translate code since it works fine with transform. Also inverted the Y axis so that Vertical movement works as it should. Finally, added joystick button 0 to Fire so that I could use that to shoot.